usageAnalytics/requirements.txt
Flask==2.1.1
Jinja2==3.0.3
Werkzeug==2.2.3
google-cloud-logging==3.6.0
google-api-python-client==2.133.0
pandas==1.5.3
matplotlib==3.6.3
numpy==1.24.3
